email_service/python/requirements.txt
nameko
importlib-metadata==4.13.0
python-dotenv==1.0.0